Refactor API into separate modules
